requirements.txt
aiohttp==3.7.4.post0
aiohttp-socks==0.6.0
aiorpcX==0.18.7
anyio==3.6.1
appdirs==1.4.3
APScheduler==3.9.1
apsw==3.28.0.post1
apt-clone==0.2.1
apturl==0.5.2
async-timeout==3.0.1
attrs==21.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.8.2
bitstring==3.1.9
blinker==1.4
Brlapi==0.7.0
cachetools==5.0.0
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
cryptography==2.8
css-parser==1.0.4
cssselect==1.1.0
cssutils==1.0.2
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.0
distro==1.4.0
dnspython==2.0.0
Electrum==4.1.5
entrypoints==0.3
evdev==1.3.0
feedparser==5.2.1
filelock==3.0.12
Flask==2.2.2
grpcio==1.16.1
h11==0.12.0
helpdev==0.7.1
html2text==2020.1.16
html5-parser==0.4.9
html5lib==1.0.1
httpcore==0.14.7
httplib2==0.14.0
httpx==0.22.0
idna==2.8
ifaddr==0.1.6
IMDbPY==6.8
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
lxml==4.5.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==2.1.1
mechanize==0.4.5
meson==0.53.2
more-itertools==4.2.0
msgpack==0.6.2
multidict==5.1.0
nemo-emblems==5.0.0
netaddr==0.7.19
netifaces==0.10.4
nose==1.3.7
oauthlib==3.1.0
onboard==1.4.1
packaging==20.3
PAM==0.4.2
pexpect==4.6.0
Pillow==7.0.0
pipenv==11.9.0
protobuf==3.17.3
psutil==5.5.1
pycairo==1.16.2
pychm==0.8.6
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.2
Pygments==2.3.1
PyGObject==3.36.0
PyICU==2.4.2
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyparted==3.11.2
PyQt5==5.14.1
PyQtWebEngine==5.14.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-magic==0.4.16
python-socks==1.2.4
python-telegram-bot==20.0a0
python-xapp==2.2.1
python-xlib==0.23
pytz==2019.3
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.26
PyYAML==5.3.1
QDarkStyle==2.8.1
qrcode==7.3
QtPy==1.11.0
regex==2019.8.19
reportlab==3.5.34
repoze.lru==0.7
requests==2.22.0
requests-file==1.4.3
requests-unixsocket==0.2.0
rfc3986==1.5.0
Routes==2.4.1
SecretStorage==2.3.1
setproctitle==1.1.10
simplejson==3.16.0
sip==4.19.21
six==1.14.0
sniffio==1.3.0
soupsieve==1.9.5
systemd-python==234
tinycss2==1.0.2
tldextract==2.2.1
tornado==6.1
typing-extensions==3.10.0.2
tzdata==2022.2
tzlocal==4.2
ubuntu-drivers-common==0.0.0
ufw==0.36
Unidecode==1.1.1
urllib3==1.25.8
virtualenv==20.0.17
virtualenv-clone==0.3.0
wadllib==1.3.3
webencodings==0.5.1
WebOb==1.8.5
Werkzeug==2.2.2
xkit==0.0.0
yarl==1.6.3
youtube-dl==2021.4.26
zeroconf==0.24.4
zipp==3.8.1
